+++ src/components/ObservedDateKey.js
@@ -4,9 +4,13 @@ const { displayDate } = require('../dates')
 const DateHtml = require('./DateHtml.js')
 const Details = require('./Details.js')
 const isSunday = require('date-fns/isSunday')
+const isMonday = require('date-fns/isMonday')
 
 const _getContext = ({ nameEn, provinces = [], date }) => {
-  if (nameEn === 'Boxing Day' && isSunday(Sugar.Date.create(date))) {
+  if (
+    nameEn === 'Boxing Day' &&
+    (isSunday(Sugar.Date.create(date)) || isMonday(Sugar.Date.create(date)))
+  ) {
     return html`<p>
       Because Christmas is observed on Monday, ${nameEn} is pushed to the following Tuesday.
     </p>`