+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.17.4] - 2022-09-30
+
+### Fixed
+
+- The little details/summary for Boxing Day 2022 claimed that it was observed on Monday, which is not the case.
+
+
 ## [3.17.3] - 2022-09-21
 
 ### Updated